JPA transactional update entity which property is modified

+++ spring-boot2-jpa/src/main/java/com/example/jpa/model/Coffee.java
@@ -18,9 +18,10 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Coffee implements Serializable {
+    
     @Id
     @GeneratedValue
-    private Long id;
+    private int id;
 
     private String name;
