+++ spring-boot2-jpa/src/main/java/com/example/jpa/repository/OrderRepository.java
@@ -4,5 +4,5 @@ package com.example.jpa.repository;
 import com.example.jpa.model.Order;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<Order, Long> {
+public interface OrderRepository extends CrudRepository<Order, Integer> {
 }
